+++ govtech_fasms/applicants/views.py
@@ -1,23 +1,15 @@
-from rest_framework import generics, permissions
-from .models import Applicant, HouseholdMember
-from .serializers import ApplicantSerializer, HouseholdMemberSerializer
+from rest_framework import viewsets, mixins
+from rest_framework.permissions import AllowAny
+from .models import Applicant
+from .serializers import ApplicantSerializer
 
-class ApplicantListCreateView(generics.ListCreateAPIView):
+class ApplicantViewSet(mixins.ListModelMixin,
+                       mixins.CreateModelMixin,
+                       mixins.RetrieveModelMixin,
+                       mixins.UpdateModelMixin,
+                       mixins.DestroyModelMixin,
+                       viewsets.GenericViewSet):
+
     queryset = Applicant.objects.all()
     serializer_class = ApplicantSerializer
-    permission_classes = [permissions.IsAuthenticated]
-
-class ApplicantDetailView(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Applicant.objects.all()
-    serializer_class = ApplicantSerializer
-    permission_classes = [permissions.IsAuthenticated]
-
-class HouseholdMemberListCreateView(generics.ListCreateAPIView):
-    queryset = HouseholdMember.objects.all()
-    serializer_class = HouseholdMemberSerializer
-    permission_classes = [permissions.IsAuthenticated]
-
-class HouseholdMemberDetailView(generics.RetrieveUpdateDestroyAPIView):
-    queryset = HouseholdMember.objects.all()
-    serializer_class = HouseholdMemberSerializer
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [AllowAny]