+++ CHANGELOG.md
@@ -11,3 +11,7 @@
 - Created serializers for each component
 - Created admin account handling using Django authenticate library
 
+## [0.0.3] - 09/12/2024
+### Added
+- Updated accounts component to allow logout
+- Updated applicants component to create unary relationship between applicants for applicant-household member relationships